+++ include/FaceEngine/Math/Vector3.h
@@ -272,7 +272,7 @@ namespace FaceEngine
 
         const std::vector<float> ToArray() const noexcept override
         {
-            return {X, Y, Z};
+            return {(float)X, (float)Y, (float)Z};
         }
     };
